requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
pyodbc==5.0.1
cryptography==41.0.7
Werkzeug==3.0.1
APScheduler==3.10.4
pythonping==1.1.4
python-dotenv==1.0.0
